+++ CHANGELOG.md
@@ -27,6 +27,11 @@
 
 - The lines of text in the output pane are better formatted
 
+### Changed
+
+- We no longer use `Math.random()` to produce random numbers but
+  `crypto.randomInt()`
+
 ## [0.1.0] - 2024-01-15
 
 - First scratch